+++ actix-web-httpauth/CHANGES.md
@@ -1,10 +1,14 @@
 # Changes
 
 ## Unreleased - 2022-xx-xx
-- `BasicAuth::user_id()` now returns a `&str`. [#249]
-- `BasicAuth::password()` now returns a `&str`. [#249]
+- `BasicAuth::user_id()` now returns `&str`. [#249]
+- `BasicAuth::password()` now returns `Option<&str>`. [#249]
+- `Basic::user_id()` now returns `&str`. [#???]
+- `Basic::password()` now returns `Option<&str>`. [#???]
+- `Bearer::token()` now returns `&str`. [#???]
 
 [#249]: https://github.com/actix/actix-extras/pull/249
+[#???]: https://github.com/actix/actix-extras/pull/???
 
 
 ## 0.7.0 - 2022-07-19